+++ packages/layout/src/components/SiderMenu/BaseMenu.tsx
@@ -373,6 +373,7 @@ const BaseMenu: React.FC<BaseMenuProps & PrivateSiderMenuProps> = (props) => {
   const {
     mode,
     className,
+    prefixCls,
     handleOpenChange,
     style,
     menuData,
@@ -528,7 +529,7 @@ const BaseMenu: React.FC<BaseMenuProps & PrivateSiderMenuProps> = (props) => {
       collapsedItemShowTitle: css`
         margin-top: 12px;
         margin-bottom: 12px;
-        .${antPrefixClassName}-pro-menu-item {
+        .${prefixCls}-menu-item {
           padding-top: 6px;
           padding-bottom: 6px;
         }
@@ -537,7 +538,7 @@ const BaseMenu: React.FC<BaseMenuProps & PrivateSiderMenuProps> = (props) => {
         .${antPrefixClassName}-menu-title-content {
           ${mode !== 'horizontal' ? 'width: 100%; line-height: 40px;' : ''}
         }
-        .${antPrefixClassName}-pro-menu-item {
+        .${prefixCls}-menu-item {
           display: flex;
           flex-direction: column;
           align-items: center;
@@ -553,10 +554,10 @@ const BaseMenu: React.FC<BaseMenuProps & PrivateSiderMenuProps> = (props) => {
         flex-direction: column;
         align-items: center;
         justify-content: center;
-        .${antPrefixClassName}-pro-menu-item {
+        .${prefixCls}-menu-item {
           padding: 0 20px;
         }
-        .${antPrefixClassName}-menu-title-content:hover {
+        .${prefixCls}-title-content:hover {
           background-color: ${itemHoverColor};
           border-radius: ${designToken.borderRadiusBase};
         }
@@ -715,7 +716,7 @@ const BaseMenu: React.FC<BaseMenuProps & PrivateSiderMenuProps> = (props) => {
       background: transparent;
     }
 
-    .${antPrefixClassName}-pro-menu-item-divider {
+    .${prefixCls}-menu-item-divider {
       &:last-child {
         display: none;
       }
@@ -726,7 +727,7 @@ const BaseMenu: React.FC<BaseMenuProps & PrivateSiderMenuProps> = (props) => {
       line-height: 20px;
     }
   `;
-  }, [antPrefixClassName, menuDesignToken.menuTextColorSecondary, props.isMobile]);
+  }, [antPrefixClassName, menuDesignToken.menuTextColorSecondary, prefixCls, props.isMobile]);
 
   if (menu?.loading) {
     return (
@@ -791,9 +792,9 @@ const BaseMenu: React.FC<BaseMenuProps & PrivateSiderMenuProps> = (props) => {
         return React.cloneElement(dom, {
           ...dom.props,
           className: cx(
-            `${antPrefixClassName}-pro-menu-submenu`,
-            stateProps?.selected && `${antPrefixClassName}-pro-menu-submenu-selected`,
-            stateProps?.open && `${antPrefixClassName}-pro-menu-submenu-open`,
+            `${prefixCls}-menu-submenu`,
+            stateProps?.selected && `${prefixCls}-menu-submenu-selected`,
+            stateProps?.open && `${prefixCls}-menu-submenu-open`,
             menuItemCssMap.subMenuItem,
             // 收起的样式
             collapsed && menuItemCssMap.collapsedSubMenuItem,
@@ -810,8 +811,8 @@ const BaseMenu: React.FC<BaseMenuProps & PrivateSiderMenuProps> = (props) => {
         return React.cloneElement(dom, {
           ...dom.props,
           className: cx(
-            `${antPrefixClassName}-pro-menu-item`,
-            stateProps?.selected && `${antPrefixClassName}-pro-menu-item-selected`,
+            `${prefixCls}-menu-item`,
+            stateProps?.selected && `${prefixCls}-menu-item-selected`,
             // 展开的样式
             menuItemCssMap.menuItem,
             // 收起的样式