+++ packages/layout/src/demos/menu-group.tsx
@@ -21,9 +21,6 @@ export default () => {
       }}
     >
       <ProLayout
-        brandBgImg={{
-          url: 'https://gw.alipayobjects.com/zos/antfincdn/tQVPs1q2X%26/yonghushenfen.png',
-        }}
         fixSiderbar
         route={{
           path: '/',