+++ packages/table/src/components/EditableTable/demos/form-linkage.tsx
@@ -205,6 +205,9 @@ export default () => {
           </ProFormDependency>
           <EditableProTable<DataSourceType>
             rowKey="id"
+            scroll={{
+              x: true,
+            }}
             controlled
             actionRef={actionRef}
             formItemProps={{