+++ packages/descriptions/CHANGELOG.md
@@ -2,6 +2,30 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.10.58](https://github.com/ant-design/pro-components/compare/@ant-design/pro-descriptions@1.10.57...@ant-design/pro-descriptions@1.10.58) (2022-03-22)
+
+**Note:** Version bump only for package @ant-design/pro-descriptions
+
+## [1.10.57](https://github.com/ant-design/pro-components/compare/@ant-design/pro-descriptions@1.10.56...@ant-design/pro-descriptions@1.10.57) (2022-03-18)
+
+**Note:** Version bump only for package @ant-design/pro-descriptions
+
+## [1.10.56](https://github.com/ant-design/pro-components/compare/@ant-design/pro-descriptions@1.10.55...@ant-design/pro-descriptions@1.10.56) (2022-03-16)
+
+**Note:** Version bump only for package @ant-design/pro-descriptions
+
+## [1.10.55](https://github.com/ant-design/pro-components/compare/@ant-design/pro-descriptions@1.10.54...@ant-design/pro-descriptions@1.10.55) (2022-03-16)
+
+**Note:** Version bump only for package @ant-design/pro-descriptions
+
+## [1.10.54](https://github.com/ant-design/pro-components/compare/@ant-design/pro-descriptions@1.10.53...@ant-design/pro-descriptions@1.10.54) (2022-03-15)
+
+**Note:** Version bump only for package @ant-design/pro-descriptions
+
+## [1.10.53](https://github.com/ant-design/pro-components/compare/@ant-design/pro-descriptions@1.10.52...@ant-design/pro-descriptions@1.10.53) (2022-03-14)
+
+**Note:** Version bump only for package @ant-design/pro-descriptions
+
 ## [1.10.52](https://github.com/ant-design/pro-components/compare/@ant-design/pro-descriptions@1.10.51...@ant-design/pro-descriptions@1.10.52) (2022-03-08)
 
 **Note:** Version bump only for package @ant-design/pro-descriptions